refactor(docs): rename theme config to style guide
