+++ themes/docs/src/pages/docs/style-guide.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+import StyleGuide from '../../components/style-guide/page'
+
+const StyleGuidePage = () => <StyleGuide />
+
+export default StyleGuidePage
