--- themes/docs/src/pages/docs/theme.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-import ThemeDocs from '../../components/theme-docs'
-
-const DocsThemePage = () => <ThemeDocs />
-
-export default DocsThemePage
